+++ app/serializers/discourse_data_explorer/small_badge_serializer.rb
@@ -1,5 +1,5 @@
 # frozen_string_literal: true
 
-class DataExplorer::SmallBadgeSerializer < ApplicationSerializer
+class DiscourseDataExplorer::SmallBadgeSerializer < ApplicationSerializer
   attributes :id, :name, :display_name, :badge_type, :description, :icon
 end
