+++ lib/discourse_data_explorer/engine.rb
@@ -1,7 +1,8 @@
 # frozen_string_literal: true
 
-module DiscourseDataExplorer
+module ::DiscourseDataExplorer
   class Engine < ::Rails::Engine
+    engine_name PLUGIN_NAME
     isolate_namespace DiscourseDataExplorer
   end
 end
