+++ lib/tasks/data_explorer.rake
@@ -2,11 +2,10 @@
 
 # rake data_explorer:list_hidden_queries
 desc "Shows a list of hidden queries"
-task("data_explorer:list_hidden_queries").clear
 task "data_explorer:list_hidden_queries" => :environment do |t|
   puts "\nHidden Queries\n\n"
 
-  hidden_queries = DataExplorer::Query.where(hidden: false)
+  hidden_queries = DiscourseDataExplorer::Query.where(hidden: false)
 
   hidden_queries.each do |query|
     puts "Name: #{query.name}"
@@ -18,11 +17,10 @@ end
 # rake data_explorer[-1]
 # rake data_explorer[1,-2,3,-4,5]
 desc "Hides one or multiple queries by ID"
-task("data_explorer").clear
 task "data_explorer" => :environment do |t, args|
   args.extras.each do |arg|
     id = arg.to_i
-    query = DataExplorer::Query.find_by(id: id)
+    query = DiscourseDataExplorer::Query.find_by(id: id)
     if query
       puts "\nFound query with id #{id}"
       query.update!(hidden: true)
@@ -37,11 +35,10 @@ end
 # rake data_explorer:unhide_query[-1]
 # rake data_explorer:unhide_query[1,-2,3,-4,5]
 desc "Unhides one or multiple queries by ID"
-task("data_explorer:unhide_query").clear
 task "data_explorer:unhide_query" => :environment do |t, args|
   args.extras.each do |arg|
     id = arg.to_i
-    query = DataExplorer::Query.find_by(id: id)
+    query = DiscourseDataExplorer::Query.find_by(id: id)
     if query
       puts "\nFound query with id #{id}"
       query.update!(hidden: false)
@@ -56,11 +53,10 @@ end
 # rake data_explorer:hard_delete[-1]
 # rake data_explorer:hard_delete[1,-2,3,-4,5]
 desc "Hard deletes one or multiple queries by ID"
-task("data_explorer:hard_delete").clear
 task "data_explorer:hard_delete" => :environment do |t, args|
   args.extras.each do |arg|
     id = arg.to_i
-    query = DataExplorer::Query.find_by(id: id)
+    query = DiscourseDataExplorer::Query.find_by(id: id)
     if query
       puts "\nFound query with id #{id}"
