+++ spec/tasks/data_explorer_spec.rb
@@ -10,7 +10,7 @@ describe "Data Explorer rake tasks" do
 
   def make_query(sql, opts = {}, group_ids = [])
     query =
-      DataExplorer::Query.create!(
+      DiscourseDataExplorer::Query.create!(
         id: opts[:id],
         name: opts[:name] || "Query number",
         description: "A description for query number",
@@ -22,19 +22,19 @@ describe "Data Explorer rake tasks" do
   end
 
   def hidden_queries
-    DataExplorer::Query.where(hidden: true).order(:id)
+    DiscourseDataExplorer::Query.where(hidden: true).order(:id)
   end
 
   describe "data_explorer" do
     it "hides a single query" do
-      DataExplorer::Query.destroy_all
+      DiscourseDataExplorer::Query.destroy_all
       make_query("SELECT 1 as value", id: 1, name: "A")
       make_query("SELECT 1 as value", id: 2, name: "B")
       # rake data_explorer[1] => hide query with ID 1
       silence_stdout { Rake::Task["data_explorer"].invoke(1) }
 
       # Soft deletion: PluginStoreRow should not be modified
-      expect(DataExplorer::Query.all.length).to eq(2)
+      expect(DiscourseDataExplorer::Query.all.length).to eq(2)
       # Array of hidden queries should have exactly 1 element
       expect(hidden_queries.length).to eq(1)
       # That one element should have the same ID as the one invoked to be hidden
@@ -42,7 +42,7 @@ describe "Data Explorer rake tasks" do
     end
 
     it "hides multiple queries" do
-      DataExplorer::Query.destroy_all
+      DiscourseDataExplorer::Query.destroy_all
       make_query("SELECT 1 as value", id: 1, name: "A")
       make_query("SELECT 1 as value", id: 2, name: "B")
       make_query("SELECT 1 as value", id: 3, name: "C")
@@ -51,7 +51,7 @@ describe "Data Explorer rake tasks" do
       silence_stdout { Rake::Task["data_explorer"].invoke(1, 2, 4) }
 
       # Soft deletion: PluginStoreRow should not be modified
-      expect(DataExplorer::Query.all.length).to eq(4)
+      expect(DiscourseDataExplorer::Query.all.length).to eq(4)
       # Array of hidden queries should have the same number of elements invoked to be hidden
       expect(hidden_queries.length).to eq(3)
       # The elements should have the same ID as the ones invoked to be hidden
@@ -62,7 +62,7 @@ describe "Data Explorer rake tasks" do
 
     context "when query does not exist in PluginStore" do
       it "should not hide the query" do
-        DataExplorer::Query.destroy_all
+        DiscourseDataExplorer::Query.destroy_all
         make_query("SELECT 1 as value", id: 1, name: "A")
         make_query("SELECT 1 as value", id: 2, name: "B")
         # rake data_explorer[3] => try to hide query with ID 3
@@ -78,14 +78,14 @@ describe "Data Explorer rake tasks" do
 
   describe "#unhide_query" do
     it "unhides a single query" do
-      DataExplorer::Query.destroy_all
+      DiscourseDataExplorer::Query.destroy_all
       make_query("SELECT 1 as value", id: 1, name: "A", hidden: true)
       make_query("SELECT 1 as value", id: 2, name: "B", hidden: true)
       # rake data_explorer:unhide_query[1] => unhide query with ID 1
       silence_stdout { Rake::Task["data_explorer:unhide_query"].invoke(1) }
 
       # Soft deletion: PluginStoreRow should not be modified
-      expect(DataExplorer::Query.all.length).to eq(2)
+      expect(DiscourseDataExplorer::Query.all.length).to eq(2)
       # Array of hidden queries should have exactly 1 element
       expect(hidden_queries.length).to eq(1)
       # There should be one remaining element that is still hidden
@@ -93,7 +93,7 @@ describe "Data Explorer rake tasks" do
     end
 
     it "unhides multiple queries" do
-      DataExplorer::Query.destroy_all
+      DiscourseDataExplorer::Query.destroy_all
       make_query("SELECT 1 as value", id: 1, name: "A", hidden: true)
       make_query("SELECT 1 as value", id: 2, name: "B", hidden: true)
       make_query("SELECT 1 as value", id: 3, name: "C", hidden: true)
@@ -102,7 +102,7 @@ describe "Data Explorer rake tasks" do
       silence_stdout { Rake::Task["data_explorer:unhide_query"].invoke(1, 2, 4) }
 
       # Soft deletion: PluginStoreRow should not be modified
-      expect(DataExplorer::Query.all.length).to eq(4)
+      expect(DiscourseDataExplorer::Query.all.length).to eq(4)
       # Array of hidden queries should have exactly 1 element
       expect(hidden_queries.length).to eq(1)
       # There should be one remaining element that is still hidden
@@ -111,7 +111,7 @@ describe "Data Explorer rake tasks" do
 
     context "when query does not exist in PluginStore" do
       it "should not unhide the query" do
-        DataExplorer::Query.destroy_all
+        DiscourseDataExplorer::Query.destroy_all
         make_query("SELECT 1 as value", id: 1, name: "A", hidden: true)
         make_query("SELECT 1 as value", id: 2, name: "B", hidden: true)
         # rake data_explorer:unhide_query[3] => try to unhide query with ID 3
@@ -127,14 +127,14 @@ describe "Data Explorer rake tasks" do
 
   describe "#hard_delete" do
     it "hard deletes a single query" do
-      DataExplorer::Query.destroy_all
+      DiscourseDataExplorer::Query.destroy_all
       make_query("SELECT 1 as value", id: 1, name: "A", hidden: true)
       make_query("SELECT 1 as value", id: 2, name: "B", hidden: true)
       # rake data_explorer:hard_delete[1] => hard delete query with ID 1
       silence_stdout { Rake::Task["data_explorer:hard_delete"].invoke(1) }
 
       # Hard deletion: query list should be shorter by 1
-      expect(DataExplorer::Query.all.length).to eq(1)
+      expect(DiscourseDataExplorer::Query.all.length).to eq(1)
       # Array of hidden queries should have exactly 1 element
       expect(hidden_queries.length).to eq(1)
       # There should be one remaining hidden element
@@ -142,7 +142,7 @@ describe "Data Explorer rake tasks" do
     end
 
     it "hard deletes multiple queries" do
-      DataExplorer::Query.destroy_all
+      DiscourseDataExplorer::Query.destroy_all
       make_query("SELECT 1 as value", id: 1, name: "A", hidden: true)
       make_query("SELECT 1 as value", id: 2, name: "B", hidden: true)
       make_query("SELECT 1 as value", id: 3, name: "C", hidden: true)
@@ -151,7 +151,7 @@ describe "Data Explorer rake tasks" do
       silence_stdout { Rake::Task["data_explorer:hard_delete"].invoke(1, 2, 4) }
 
       # Hard deletion: query list should be shorter by 3
-      expect(DataExplorer::Query.all.length).to eq(1)
+      expect(DiscourseDataExplorer::Query.all.length).to eq(1)
       # Array of hidden queries should have exactly 1 element
       expect(hidden_queries.length).to eq(1)
       # There should be one remaining hidden element
@@ -160,7 +160,7 @@ describe "Data Explorer rake tasks" do
 
     context "when query does not exist in PluginStore" do
       it "should not hard delete the query" do
-        DataExplorer::Query.destroy_all
+        DiscourseDataExplorer::Query.destroy_all
         make_query("SELECT 1 as value", id: 1, name: "A", hidden: true)
         make_query("SELECT 1 as value", id: 2, name: "B", hidden: true)
         # rake data_explorer:hard_delete[3] => try to hard delete query with ID 3
@@ -175,13 +175,13 @@ describe "Data Explorer rake tasks" do
 
     context "when query is not hidden" do
       it "should not hard delete the query" do
-        DataExplorer::Query.destroy_all
+        DiscourseDataExplorer::Query.destroy_all
         make_query("SELECT 1 as value", id: 1, name: "A")
         # rake data_explorer:hard_delete[1] => try to hard delete query with ID 1
         silence_stdout { Rake::Task["data_explorer:hard_delete"].invoke(1) }
 
         # List of queries shouldn't change
-        expect(DataExplorer::Query.all.length).to eq(1)
+        expect(DiscourseDataExplorer::Query.all.length).to eq(1)
       end
     end
   end
